Convert the NodeFilterPropertiesPipe to a method on the sidebar component
